Merge brach 'main' of https://github.com/dJuKaAaA/ISS-Tim-7 into member1-kt2

+++ tim7iss/src/main/java/com/tim7/iss/tim7iss/services/LocationService.java
@@ -19,7 +19,7 @@ public class LocationService {
         return locationRepository.findById(id).orElse(null);
     }
 
-    public Location getByLongitudeAndLatitude(double longitude, double latitude) {
+    public Location getByLongitudeAndLatitude(float longitude, float latitude) {
         return locationRepository.findByLongitudeAndLatitude(longitude, latitude).orElse(null);
     }
